Move {show,hide}{Gridlines,DirectionArrows} into Main window
