+++ src/OpenLoco/src/Ui/WindowManager.cpp
@@ -1,6 +1,5 @@
 #include "WindowManager.h"
 #include "Audio/Audio.h"
-#include "Config.h"
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
@@ -2234,87 +2233,6 @@ namespace OpenLoco::Ui::WindowManager
                 continue;
 
             windowDraw(&rt, w, rect);
-        }
-    }
-}
-
-namespace OpenLoco::Ui::Windows
-{
-    static loco_global<int8_t, 0x00F2533F> _gridlinesState;
-    static loco_global<uint8_t, 0x0112C2E1> _directionArrowsState;
-
-    // 0x00468FD3
-    void showGridlines()
-    {
-        if (!_gridlinesState)
-        {
-            auto window = WindowManager::getMainWindow();
-            if (window != nullptr)
-            {
-                if (!window->viewports[0]->hasFlags(ViewportFlags::gridlines_on_landscape))
-                {
-                    window->invalidate();
-                }
-                window->viewports[0]->flags |= ViewportFlags::gridlines_on_landscape;
-            }
-        }
-        _gridlinesState++;
-    }
-
-    // 0x00468FFE
-    void hideGridlines()
-    {
-        _gridlinesState--;
-        if (!_gridlinesState)
-        {
-            if (!Config::get().hasFlags(Config::Flags::gridlinesOnLandscape))
-            {
-                auto window = WindowManager::getMainWindow();
-                if (window != nullptr)
-                {
-                    if (window->viewports[0]->hasFlags(ViewportFlags::gridlines_on_landscape))
-                    {
-                        window->invalidate();
-                    }
-                    window->viewports[0]->flags &= ~ViewportFlags::gridlines_on_landscape;
-                }
-            }
-        }
-    }
-
-    // 0x004793C4
-    void showDirectionArrows()
-    {
-        if (!_directionArrowsState)
-        {
-            auto mainWindow = WindowManager::getMainWindow();
-            if (mainWindow != nullptr)
-            {
-                if (!mainWindow->viewports[0]->hasFlags(ViewportFlags::one_way_direction_arrows))
-                {
-                    mainWindow->viewports[0]->flags |= ViewportFlags::one_way_direction_arrows;
-                    mainWindow->invalidate();
-                }
-            }
-        }
-        _directionArrowsState++;
-    }
-
-    // 0x004793EF
-    void hideDirectionArrows()
-    {
-        _directionArrowsState--;
-        if (!_directionArrowsState)
-        {
-            auto mainWindow = WindowManager::getMainWindow();
-            if (mainWindow != nullptr)
-            {
-                if (mainWindow->viewports[0]->hasFlags(ViewportFlags::one_way_direction_arrows))
-                {
-                    mainWindow->viewports[0]->flags &= ~ViewportFlags::one_way_direction_arrows;
-                    mainWindow->invalidate();
-                }
-            }
         }
     }
 }
